+++ python/spglib/__init__.py
@@ -49,6 +49,7 @@ from .spglib import (  # noqa: F401
     get_ir_reciprocal_mesh,
     get_layergroup,
     get_magnetic_spacegroup_type,
+    get_magnetic_spacegroup_type_from_symmetry,
     get_magnetic_symmetry,
     get_magnetic_symmetry_dataset,
     get_magnetic_symmetry_from_database,